[SPARK-56875][SQL] Refactor time/session window resolution code to be reused in single-pass Analyzer
